Ten conflicted files, resolved per the verified base-class contracts:

- GLA/GatedDeltaNet/Griffin/Hawk/RecurrentGemma/Finch Options copy ctors:
  take master. The base chain already copies both members this branch
  re-copied -- ModelOptions copies Seed (and holds the ArgumentNullException
  guard), NeuralNetworkOptions copies EncoderLayerCount -- so the branch's
  lines were redundant and its `if (other is null) throw` unreachable,
  because `: base(other)` runs first.
- FinchOptions second hunk: take master's EnableGradientClipping /
  MaxGradientNorm copies AND restore master's declarations for them. The
  merged property block kept this branch's Beta1/Beta2 shape and dropped
  master's clipping knobs, so master's assignments alone would not compile.
  The branch's other copies (MaxGradNorm, VocabSize, ModelDimension,
  NumLayers, NumHeads, MaxSequenceLength) are already done by
  ModelHyperparameterOptions and SequenceModelOptions.
- FinchLanguageModel: keep this branch's _learningRate/_optimizer and its
  AdamW construction from the published recipe. FinchOptions declares every
  member it reads, and master simply does not have this feature.
- OptionsSurfaceRatchetTests: take master (it adds a real assertion and the
  GapReport test) and add the `using Xunit.Abstractions;` master's
  CapturingOutput needs -- the file imported only Xunit.
- PrivBayesDifferentialPrivacyTests: take master; it asserts the setter
  itself rejects the degenerate split.
- AudioHyperparameterOptions.cs: keep master's file rather than this
  branch's deletion. Master's SharedOptionsDocumentationContractTests does
  typeof(AudioHyperparameterOptions), so the deletion would not compile.
  This one is a genuine intent clash and deserves the author's review.

+++ src/Preprocessing/FeatureSelection/Filter/Statistical/WilcoxonSelector.cs
@@ -89,10 +89,10 @@ public class WilcoxonSelector<T> : TransformerBase<T, Matrix<T>, Matrix<T>>
             double r0 = class0.Sum(i => ranked[i]);
 
             // U statistic for class 0
-            double u0 = r0 - (n0 * (n0 + 1)) / 2.0;
+            double u0 = r0 - ((double)n0 * (n0 + 1)) / 2.0;
 
             // Normalize U to [0, 1]
-            double maxU = n0 * n1;
+            double maxU = (double)n0 * n1;
             _uStatistics[j] = maxU > 0 ? Math.Abs(u0 - maxU / 2) / (maxU / 2) : 0;
         }
